+++ src/core/TimePicker/__tests__/TimerPicker.test.tsx
@@ -1,16 +1,16 @@
 import { render } from '@testing-library/react';
 import React from 'react';
 import { KuberaThemeProvider } from '../../../theme';
-import { BaseTimePicker } from '../BaseTimePicker';
+import { TimePicker } from '../TimePicker';
 
 describe('Timepicker component', () => {
   it('Renders', () => {
     const { getByTestId } = render(
       <KuberaThemeProvider platform="kubera-chaos">
-        <BaseTimePicker onChange={() => console.log('timepicker')} />
+        <TimePicker onChange={() => console.log('Time Changed')} />
       </KuberaThemeProvider>
     );
 
-    expect(getByTestId('timepicker')).toBeTruthy();
+    expect(getByTestId('Time Changed')).toBeTruthy();
   });
 });
