+++ src/core/TimePicker/base.ts
@@ -1,0 +1,3 @@
+import { KeyboardTimePickerProps } from '@material-ui/pickers';
+
+export type BaseTimePickerProps = Omit<KeyboardTimePickerProps, 'value'>;
