+++ src/core/TimePicker/index.ts
@@ -1,2 +1,1 @@
-export * from './Button';
 export * from './TimePicker';
