--- src/core/Timepicker/base.ts
@@ -1,3 +1,0 @@
-import { TimePickerProps } from '@material-ui/pickers';
-
-export type BaseTimePickerProps = Omit<TimePickerProps, 'value'>;
