--- src/core/Timepicker/index.ts
@@ -1,1 +1,0 @@
-export * from './BaseTimePicker';
