--- src/core/Timepicker/style.ts
@@ -1,5 +1,0 @@
-import { makeStyles } from '@material-ui/core';
-
-const useStyles = makeStyles(() => ({}));
-
-export { useStyles };
